// frontend/src/types/models.ts
export type Rank = 'JUNIOR' | 'REGULAR' | 'SENIOR' | 'MASTER' | 'LEGEND'